requirements.txt
asgiref==3.8.1
Django==5.2.3
djangorestframework==3.16.0
drf-yasg==1.21.10
inflection==0.5.1
packaging==25.0
pillow==11.2.1
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
uritemplate==4.2.0

boto3~=1.39.0
botocore~=1.39.0